Added requirements for gps, improved permissions requests, added notifications

+++ app/src/main/java/todo_database/Shop.kt
@@ -4,7 +4,7 @@ class Shop(
         var id: String?,
         var name: String,
         var description: String,
-        var radius: Double,
+        var radius: Float,
         var coordinates: String) {
-    constructor() : this(null, "", "", -1.0, "")
+    constructor() : this(null, "", "", -1.0f, "")
 }
